Moved source code into src folder.
- created src folder for cleaner structure
- fixed import paths
