+++ src/image_quality_analyzer.py
